fixed loading crop frames, more feedback on encoding success, encoding dialog doesn't jump around anymore

+++ CuteVideoEditor.Core/Models/SerializationModels.cs
@@ -18,7 +18,7 @@ public class RectSerializationModel
 
 public class CropFrameEntrySerializationModel
 {
-    public int FrameNumber { get; set; }
+    public long FrameNumber { get; set; }
     public RectSerializationModel? CropRectangle { get; set; }
 }
